web and degin responsiveness fixed
