--- app/global-error.jsx
@@ -1,13 +1,0 @@
-"use client";
-import React from "react";
-
-export default function globalError() {
-  return (
-    <html>
-      <body>
-        <h2>Something went wrong!</h2>
-        <button onClick={() => reset()}>Try again</button>
-      </body>
-    </html>
-  );
-}
